Simulation performance enhancements.
Replacing an einsum with an equivalent operation that uses BLAS has
minor performance advantages on my dev PC, but makes a big difference on
some other computers. STDP code now clips in-place, avoiding some
unnecessary allocations.

Also changed the mechanics of the refractory period in a LIF neuron to
match the paper I'm trying to replicate: during the refractory period,
the neuron continues to integrate, but can't fire.