fix(randomization): just warn, not error, for dt
Previously, if you tried to randomize a SpikeData and you ended up
aliasing two spikes into the same bin, it would fail with a ValueError.
But some datasets are ill-formed and include multiple spikes at the same
time, so it's impossible to use this method.

Now only a warning is produced, so it's up to the user to consider
whether that aliasing is enough of a problem.