fix(randomization): sporadic ValueError
In `randomize_raster()`, when new units are being selected to provide
the spikes for a given bin, previously sampling without replacement was
used. This meant that sometimes there would not be enough units left to
create a random sample, resulting in a ValueError.

This commit replaces this approach with a new method called
`best_effort_sample()`, which attempts to approximate sampling without
replacement as closely as possible given the pigeonhole principle etc.
The maximum number of repeats in the returned sample should be the
minimum number possible given what was requested.

Fix #13