fix(randomization): randomized spikes in wrong bin
The reason spike counts were often wrong after randomization wasn't
actually in `randomize_raster()`, it was the process of translating a
correctly randomized raster back into a `SpikeData`. Spikes were being
placed directly on the boundary between the raster bins, so converting
*back* to a raster would assign the spikes to the wrong bin.

This patch puts all the spikes in the reconverted `SpikeData` right in
the middle of the raster bins, so `SpikeData.randomized()` now preserves
the invariant it promises to.

Also rename the argument `dt` of `SpikeData.randomized()` to `bin_size`
to clarify that it is being used for binning.

Fix #13