build(utils): move dependencies to minimal
@davidparks21 pointed out that utils are supposed to be usable even with
only the "minimal" dependencies, and OKed moving `joblib`, `awswrangler`,
and `smart_open` all to "minimal". This also allows removing
`awswrangler` and `smart_open` from the "analysis" dependencies, so
their versions only have to be specified in one place.

@DailyDreaming suggested putting an upper limit on the joblib dependency
as well in case there are breaking changes in the next major version.