Fix first failing load_spike_data() test
The file being loaded in this test has "cluster_KSLabel.tsv" where the
cluster groups are in a column "KSLabel" rather than "group" as
previously assumed.

This commit handles this case by renaming that column, as well as
reducing code duplication and handling another TSV name that is present
in some Phy zips.

Only a partial solution to #117 (second test still failing).