Partial revert of fdbb47f
The `@fails_on_windows` decorator which that commit tried to introduce
doesn't work correctly in combination with other `unittest` decorators
on Windows, despite working exactly as intended on Linux if I switch
which platform is expected to fail. I assume that means it's some
esoteric interaction I don't have the Windows experience to debug.