Fix CI (#118)
* Finally use spikedata from PyPI

* Fix first failing load_spike_data() test

The file being loaded in this test has "cluster_KSLabel.tsv" where the
cluster groups are in a column "KSLabel" rather than "group" as
previously assumed.

This commit handles this case by renaming that column, as well as
reducing code duplication and handling another TSV name that is present
in some Phy zips.

Only a partial solution to #117 (second test still failing).

* Fix second failing load_spike_data() test

In this case, the problem was that some labeled clusters were not
actually assigned to any spike in the actual recording. This caused
KeyError when trying to find a template for that spike while
constructing the neuron attributes.

Fix #117

* Correct authentication command in docstrings

Just changing `python -m braingeneers.iot.messaging authenticate`
to `python -m braingeneers.iot.authenticate`.

* Skip messaging tests if JWT missing

* Include Python 3.12 in testing

* Stop trying to upload coverage reports

The upload action was never successful due to branch protection. These
reports would be useless anyway since coverage is low and many tests are
skipped.

* Add "All Pass" step to CI for branch protection

* Update test skips which seem to work now

Some tests were being skipped but actually do work on my machine, so
hopefully they'll work in CI too.

* Skip another test that breaks on Windows

* Fix memoize_s3 environment variable behavior

One test failed on Windows because it was overriding the entire
environment instead of simply clearing the value of `$S3_USER`.
This patch also treats blank as unset, which is the correct behavior,
and allows writing this test to only change that one value.

* Clean up test skipping and TODOs

This commit switches most skipped tests into `unittest.expectedFail`
so that they will still be checked, and commits that fix them can't
forget to turn them back on.

It also adds a decorator `@fails_on_windows` to use instead of a
conditional skip for tests that don't work on Windows for the same
reason.

Finally, TODO comments have been converted into GitHub issues.

* Partial revert of fdbb47f

The `@fails_on_windows` decorator which that commit tried to introduce
doesn't work correctly in combination with other `unittest` decorators
on Windows, despite working exactly as intended on Linux if I switch
which platform is expected to fail. I assume that means it's some
esoteric interaction I don't have the Windows experience to debug.

* Stop continuing on error in CI!

This causes "All Checks Passed" even when they didn't. The setting we
actually wanted is `fail-fast: false`, which makes the whole matrix
execute even if one fails.

* Make `@memoize` fall back to local when credentials are missing

Fix #38

* Fix memoize_s3 on Windows

Trivial fix: the only problem was that `_open_file()` wasn't calling
`normalize_location()`.

---------

Co-authored-by: David Parks <davidparks21@gmail.com>